requirements.txt
Flask
torch
torchvision
numpy
matplotlib
pillow
numpy